backend/requirements.txt
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
sqlmodel==0.0.14
asyncpg==0.28.0